backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
Pillow==10.3.0
requests==2.31.0
python-dateutil==2.8.2
channels==4.1.0
daphne==4.1.2
